be/requirements.txt
flask==3.0.0
flask-cors==4.0.0
zeroconf==0.132.2
pyyaml==6.0.1
jsonschema==4.20.0
